Refactor: Seperate Model
- 모델 분리 및 정리
- PokemonInfo -> PokemonModel로 이름 변경
