+++ PokemonDex/TodaysPokemonViewController.swift
@@ -7,428 +7,6 @@
 
 import UIKit
 
-struct PokemonInfo: Codable {
-    var abilities: [PokemonAbility]
-    var baseExperience: Int
-    var cries: PokemonCries
-    var forms: [PokemonForm]
-    var gameIndices: [PokemonGameIndex]
-    var height: Int
-    var id: Int
-    var isDefault: Bool
-    var locationAreaEncounters: String
-    var moves: [PokemonMove]
-    var name: String
-    var order: Int
-    var species: PokemonSpecies
-    var sprites: PokemonSprites
-    var stats: [PokemonStat]
-    var types: [PokemonType]
-    var weight: Int
-
-    enum CodingKeys: String, CodingKey {
-        case abilities
-        case baseExperience = "base_experience"
-        case cries
-        case forms
-        case gameIndices = "game_indices"
-        case height
-        case id
-        case isDefault = "is_default"
-        case locationAreaEncounters = "location_area_encounters"
-        case moves
-        case name
-        case order
-        case species
-        case sprites
-        case stats
-        case types
-        case weight
-    }
-}
-
-struct PokemonAbility: Codable {
-    var ability: PokemonAbilityComponent
-    var isHidden: Bool
-    var slot: Int
-
-    enum CodingKeys: String, CodingKey {
-        case ability
-        case isHidden = "is_hidden"
-        case slot
-    }
-}
-
-struct PokemonAbilityComponent: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonCries: Codable {
-    var latest: String
-    var legacy: String?
-}
-
-struct PokemonForm: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonGameIndex: Codable {
-    var gameIndex: Int
-    var version: PokemonGameIndexVersion
-
-    enum CodingKeys: String, CodingKey {
-        case gameIndex = "game_index"
-        case version
-    }
-}
-
-struct PokemonGameIndexVersion: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonMove: Codable {
-    var move: PokemonMoveComponent
-    var versionGroupDetails: [PokemonMoveVersionGroupDetail]
-
-    enum CodingKeys: String, CodingKey {
-        case move
-        case versionGroupDetails = "version_group_details"
-    }
-}
-
-struct PokemonMoveComponent: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonMoveVersionGroupDetail: Codable {
-    var levelLearnedAt: Int
-    var moveLearnMethod: PokemonMoveLearnMethod
-    var versionGroup: PokemonMoveVersionGroup
-
-    enum CodingKeys: String, CodingKey {
-        case levelLearnedAt = "level_learned_at"
-        case moveLearnMethod = "move_learn_method"
-        case versionGroup = "version_group"
-    }
-}
-
-struct PokemonMoveLearnMethod: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonMoveVersionGroup: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonSpecies: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonSprites: Codable {
-    var backDefault: String?
-    var backFemale: String?
-    var backShiny: String?
-    var backShinyFemale: String?
-    var frontDefault: String?
-    var frontFemale: String?
-    var frontShiny: String?
-    var frontShinyFemale: String?
-    var other: PokemonSpritesOther
-    var versions: PokemonSpritesVersions
-
-    enum CodingKeys: String, CodingKey {
-        case backDefault = "back_default"
-        case backFemale = "back_female"
-        case backShiny = "back_shiny"
-        case backShinyFemale = "back_shiny_female"
-        case frontDefault = "front_default"
-        case frontFemale = "front_female"
-        case frontShiny = "front_shiny"
-        case frontShinyFemale = "front_shiny_female"
-        case other
-        case versions
-    }
-}
-
-struct PokemonSpritesComponent: Codable {
-    var backDefault: String?
-    var backGray: String?
-    var backTransparent: String?
-    var backFemale: String?
-    var backShiny: String?
-    var backShinyTransparent: String?
-    var backShinyFemale: String?
-    var frontDefault: String?
-    var frontGray: String?
-    var frontTransparent: String?
-    var frontFemale: String?
-    var frontShiny: String?
-    var frontShinyTransparent: String?
-    var frontShinyFemale: String?
-    var animated: PokemonAnimatedSpritesComponent?
-
-    enum CodingKeys: String, CodingKey {
-        case backDefault = "back_default"
-        case backGray = "back_gray"
-        case backTransparent = "back_transparent"
-        case backFemale = "back_female"
-        case backShiny = "back_shiny"
-        case backShinyTransparent = "back_shiny_transparent"
-        case backShinyFemale = "back_shiny_female"
-        case frontDefault = "front_default"
-        case frontGray = "front_gray"
-        case frontTransparent = "front_transparent"
-        case frontFemale = "front_female"
-        case frontShiny = "front_shiny"
-        case frontShinyTransparent = "front_shiny_transparent"
-        case frontShinyFemale = "front_shiny_female"
-        case animated
-    }
-}
-
-struct PokemonAnimatedSpritesComponent: Codable {
-    var backDefault: String?
-    var backGray: String?
-    var backTransparent: String?
-    var backFemale: String?
-    var backShiny: String?
-    var backShinyTransparent: String?
-    var backShinyFemale: String?
-    var frontDefault: String?
-    var frontGray: String?
-    var frontTransparent: String?
-    var frontFemale: String?
-    var frontShiny: String?
-    var frontShinyTransparent: String?
-    var frontShinyFemale: String?
-
-    enum CodingKeys: String, CodingKey {
-        case backDefault = "back_default"
-        case backGray = "back_gray"
-        case backTransparent = "back_transparent"
-        case backFemale = "back_female"
-        case backShiny = "back_shiny"
-        case backShinyTransparent = "back_shiny_transparent"
-        case backShinyFemale = "back_shiny_female"
-        case frontDefault = "front_default"
-        case frontGray = "front_gray"
-        case frontTransparent = "front_transparent"
-        case frontFemale = "front_female"
-        case frontShiny = "front_shiny"
-        case frontShinyTransparent = "front_shiny_transparent"
-        case frontShinyFemale = "front_shiny_female"
-    }
-}
-
-struct PokemonSpritesOther: Codable {
-    var dreamWorld: PokemonSpritesComponent
-    var home: PokemonSpritesComponent
-    var officialArtwork: PokemonSpritesComponent
-    var showdown: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case dreamWorld = "dream_world"
-        case home
-        case officialArtwork = "official-artwork"
-        case showdown
-    }
-}
-
-struct PokemonSpritesVersions: Codable {
-    var generationI: PokemonSpritesGenerationI
-    var generationII: PokemonSpritesGenerationII
-    var generationIII: PokemonSpritesGenerationIII
-    var generationIV: PokemonSpritesGenerationIV
-    var generationV: PokemonSpritesGenerationV
-    var generationVI: PokemonSpritesGenerationVI
-    var generationVII: PokemonSpritesGenerationVII
-    var generationVIII: PokemonSpritesGenerationVIII
-
-    enum CodingKeys: String, CodingKey {
-        case generationI = "generation-i"
-        case generationII = "generation-ii"
-        case generationIII = "generation-iii"
-        case generationIV = "generation-iv"
-        case generationV = "generation-v"
-        case generationVI = "generation-vi"
-        case generationVII = "generation-vii"
-        case generationVIII = "generation-viii"
-    }
-}
-
-struct PokemonSpritesGenerationI: Codable {
-    var redBlue: PokemonSpritesComponent
-    var yellow: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case redBlue = "red-blue"
-        case yellow
-    }
-}
-
-struct PokemonSpritesGenerationII: Codable {
-    var crystal: PokemonSpritesComponent
-    var gold: PokemonSpritesComponent
-    var silver: PokemonSpritesComponent
-}
-
-struct PokemonSpritesGenerationIII: Codable {
-    var emerald: PokemonSpritesComponent
-    var fireredLeafgreen: PokemonSpritesComponent
-    var rubySapphire: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case emerald
-        case fireredLeafgreen = "firered-leafgreen"
-        case rubySapphire = "ruby-sapphire"
-    }
-}
-
-struct PokemonSpritesGenerationIV: Codable {
-    var diamondPearl: PokemonSpritesComponent
-    var heartgoldSoulsilver: PokemonSpritesComponent
-    var platinum: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case diamondPearl = "diamond-pearl"
-        case heartgoldSoulsilver = "heartgold-soulsilver"
-        case platinum
-    }
-}
-
-struct PokemonSpritesGenerationV: Codable {
-    var blackWhite: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case blackWhite = "black-white"
-    }
-}
-
-struct PokemonSpritesGenerationVI: Codable {
-    var omegarubyAlphasapphire: PokemonSpritesComponent
-    var xy: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case omegarubyAlphasapphire = "omegaruby-alphasapphire"
-        case xy = "x-y"
-    }
-}
-
-struct PokemonSpritesGenerationVII: Codable {
-    var icons: PokemonSpritesComponent
-    var ultraSunUltraMoon: PokemonSpritesComponent
-
-    enum CodingKeys: String, CodingKey {
-        case icons
-        case ultraSunUltraMoon = "ultra-sun-ultra-moon"
-    }
-}
-
-struct PokemonSpritesGenerationVIII: Codable {
-    var icons: PokemonSpritesComponent
-}
-
-struct PokemonStat: Codable {
-    var baseStat: Int
-    var effort: Int
-    var stat: PokemonStatComponent
-
-    enum CodingKeys: String, CodingKey {
-        case baseStat = "base_stat"
-        case effort
-        case stat
-    }
-}
-
-struct PokemonStatComponent: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonType: Codable {
-    var slot: Int
-    var type: PokemonTypeComponent
-}
-
-struct PokemonTypeComponent: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokemonSpeciesInfo: Codable {
-    var flavorTextEntries: [PokedexFlavorText]
-    var genera: [PokedexGenus]
-    var id: Int
-    var name: String
-    var names: [PokedexName]
-    var order: Int
-    var pokedexNumbers: [PokedexNumber]
-
-    enum CodingKeys: String, CodingKey {
-        case flavorTextEntries = "flavor_text_entries"
-        case genera
-        case id
-        case name
-        case names
-        case order
-        case pokedexNumbers = "pokedex_numbers"
-    }
-}
-
-struct PokedexFlavorText: Codable {
-    var flavorText: String
-    var language: PokedexLanguage
-    var version: PokedexVersion
-
-    enum CodingKeys: String, CodingKey {
-        case flavorText = "flavor_text"
-        case language
-        case version
-    }
-}
-
-struct PokedexLanguage: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokedexVersion: Codable {
-    var name: String
-    var url: String
-}
-
-struct PokedexGenus: Codable {
-    var genus: String
-    var language: PokedexLanguage
-}
-
-struct PokedexName: Codable {
-    var language: PokedexLanguage
-    var name: String
-}
-
-struct PokedexNumber: Codable {
-    var entryNumber: Int
-    var pokedex: PokedexNumberInfo
-
-    enum CodingKeys: String, CodingKey {
-        case entryNumber = "entry_number"
-        case pokedex
-    }
-}
-
-struct PokedexNumberInfo: Codable {
-    var name: String
-    var url: String
-}
-
 class TodaysPokemonViewController: UIViewController {
 
     private let todaysPokemonBackground: UIView = {
@@ -696,7 +274,7 @@ class TodaysPokemonViewController: UIViewController {
                 return
             }
 
-            guard let json = try? JSONDecoder().decode(PokemonInfo.self, from: data) else {
+            guard let json = try? JSONDecoder().decode(PokemonModel.self, from: data) else {
                 print("Error: Data Decoding error")
                 return
             }
@@ -954,7 +532,7 @@ class TodaysPokemonViewController: UIViewController {
                         return
                     }
 
-                    guard let json = try? JSONDecoder().decode(PokemonInfo.self, from: data) else {
+                    guard let json = try? JSONDecoder().decode(PokemonModel.self, from: data) else {
                         print("Error: Data Decoding error")
                         return
                     }
