+++ PokemonDex/UtilityModel.swift
@@ -1,0 +1,13 @@
+//
+//  UtilityModel.swift
+//  PokemonDex
+//
+//  Created by 김민택 on 8/10/24.
+//
+
+import Foundation
+
+struct NamedAPIResource: Codable {
+    var name: String
+    var url: String
+}
